feat: consolidate mithril commands
