+++ crates/amaru-mithril/Cargo.toml
@@ -19,15 +19,13 @@ categories.workspace = true
 # External dependencies ───────────────────────────────────────────────────────┐
 anyhow.workspace = true
 async-trait.workspace = true
-flate2.workspace = true
-indicatif.workspace = true
 mithril-client.workspace = true
-tar.workspace = true
 tokio = { workspace = true, features = ["sync"] }
 tracing.workspace = true
 
 # Internal dependencies ───────────────────────────────────────────────────────┐
 amaru-kernel.workspace = true
+amaru-progress-bar.workspace = true
 
 [target.'cfg(all(not(target_family="wasm"), not(target_arch="riscv32")))'.dependencies]
 amaru-deps.workspace = true
