+++ crates/amaru/src/bin/amaru/cmd/mod.rs
@@ -17,6 +17,8 @@ use std::{ops::Deref, str::FromStr};
 use amaru_kernel::{HeaderHash, Point};
 
 pub(crate) mod dev;
+#[cfg(feature = "mithril")]
+pub(crate) mod mithril;
 pub(crate) mod node;
 pub(crate) mod shell_completions;
 pub(crate) mod snapshot;
